requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astor==0.8.1
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
gast==0.2.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.2
opencv-python==4.4.0.44
opt-einsum==3.3.0
pathspec==0.8.0
Pillow==7.2.0
protobuf==3.13.0
python-dotenv==0.14.0
regex==2020.10.11
requests==2.24.0
scipy==1.5.2
six==1.15.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
